implement conversion to arbitrary prefixes
This commit adds an `prefix` argument to the `Address.cash_address` method.
If not specified, it defaults to the current prefix if the object was
instantiated  from string via a CashAddr, or to the "bitcoincash:" prefix if
it was instantiated from string with a legacy address.

Test plan:

python3 -m tests.test