support uppercase prefixes
If the users specifies an uppercase prefix, the output should be entirely
uppercase. Mixed-case prefixes are invalid.

Note that the case is not recorded in the instance, when creating an Address
from string. If the prefix is not explicitely specified and uppercase, the
output from `Address.cash_address()` will always be lowercase.

Test plan:
python3 -m tests.test